Adds RatableAmount - Loan and investments extends RatableAmount
