+++ src/Loan.php
@@ -1,0 +1,9 @@
+<?php
+
+
+namespace Soisy;
+
+
+class Loan extends RatableAmount
+{
+}
